refactor[#293] : Date 객체를 입력값으로 받아 yyyy.mm.dd 형태인 문자열을 반환하는 함수 위치 이동
