+++ src/features/marking/ui/markingItem.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import { API_BASE_URL } from "@/shared/constants";
+import { formatDateToYearMonthDay } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
 import { Button } from "@/shared/ui/button";
 import { DividerLine } from "@/shared/ui/divider";
@@ -21,16 +22,6 @@ import {
   useDeleteLikeMarking,
 } from "../api";
 import { useDeleteMarking } from "../api";
-
-function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0"); // 월은 0부터 시작하므로 1을 더함
-  const day = String(date.getDate()).padStart(2, "0");
-
-  return `${year}.${month}.${day}`;
-}
 
 type MarkingItemProps = {
   onRegionClick: () => void;
@@ -218,7 +209,7 @@ export const MarkingItem = ({
 
       <p className="text-grey-700 body-2 text-overflow">{content}</p>
 
-      <p className="body-3 text-grey-500">{formatDate(regDt)}</p>
+      <p className="body-3 text-grey-500">{formatDateToYearMonthDay(regDt)}</p>
     </li>
   );
 };
